+++ fastapi_sqla/async_sqla.py
@@ -1,19 +1,16 @@
-import math
 import os
-from collections.abc import AsyncGenerator, Awaitable, Callable
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Annotated, Iterator, Optional, Union, cast
+from typing import Annotated
 
 import structlog
-from fastapi import Depends, Query, Request
+from fastapi import Depends, Request
 from sqlalchemy import text
 from sqlalchemy.ext.asyncio import AsyncEngine
 from sqlalchemy.ext.asyncio import AsyncSession as SqlaAsyncSession
 from sqlalchemy.orm.session import sessionmaker
-from sqlalchemy.sql import Select, func, select
 
 from fastapi_sqla import aws_aurora_support, aws_rds_iam_support
-from fastapi_sqla.models import Page
 from fastapi_sqla.sqla import _DEFAULT_SESSION_KEY, Base, new_engine
 
 logger = structlog.get_logger(__name__)
@@ -152,84 +149,4 @@ class AsyncSessionDependency:
             )
 
 
-QueryCountDependency = Callable[..., Awaitable[int]]
-PaginateSignature = Callable[[Select, Optional[bool]], Awaitable[Page]]
-DefaultDependency = Callable[[SqlaAsyncSession, int, int], PaginateSignature]
-WithQueryCountDependency = Callable[
-    [SqlaAsyncSession, int, int, int], PaginateSignature
-]
-PaginateDependency = Union[DefaultDependency, WithQueryCountDependency]
-
-
-async def default_query_count(session: SqlaAsyncSession, query: Select) -> int:
-    result = await session.execute(select(func.count()).select_from(query.subquery()))
-    return cast(int, result.scalar())
-
-
-async def paginate_query(
-    query: Select,
-    session: SqlaAsyncSession,
-    total_items: int,
-    offset: int,
-    limit: int,
-    *,
-    scalars: bool = True,
-) -> Page:
-    total_pages = math.ceil(total_items / limit)
-    page_number = offset / limit + 1
-    query = query.offset(offset).limit(limit)
-    result = await session.execute(query)
-    data = iter(
-        cast(Iterator, result.unique().scalars() if scalars else result.mappings())
-    )
-    return Page(
-        data=data,
-        meta={
-            "offset": offset,
-            "total_items": total_items,
-            "total_pages": total_pages,
-            "page_number": page_number,
-        },
-    )
-
-
-def AsyncPagination(
-    session_key: str = _DEFAULT_SESSION_KEY,
-    min_page_size: int = 10,
-    max_page_size: int = 100,
-    query_count: Union[QueryCountDependency, None] = None,
-) -> PaginateDependency:
-    def default_dependency(
-        session: SqlaAsyncSession = Depends(AsyncSessionDependency(key=session_key)),
-        offset: int = Query(0, ge=0),
-        limit: int = Query(min_page_size, ge=1, le=max_page_size),
-    ) -> PaginateSignature:
-        async def paginate(query: Select, scalars=True) -> Page:
-            total_items = await default_query_count(session, query)
-            return await paginate_query(
-                query, session, total_items, offset, limit, scalars=scalars
-            )
-
-        return paginate
-
-    def with_query_count_dependency(
-        session: SqlaAsyncSession = Depends(AsyncSessionDependency(key=session_key)),
-        offset: int = Query(0, ge=0),
-        limit: int = Query(min_page_size, ge=1, le=max_page_size),
-        total_items: int = Depends(query_count),
-    ):
-        async def paginate(query: Select, scalars=True) -> Page:
-            return await paginate_query(
-                query, session, total_items, offset, limit, scalars=scalars
-            )
-
-        return paginate
-
-    if query_count:
-        return with_query_count_dependency
-    else:
-        return default_dependency
-
-
-AsyncPaginate = Annotated[PaginateDependency, Depends(AsyncPagination())]
 AsyncSession = Annotated[SqlaAsyncSession, Depends(AsyncSessionDependency())]
